feat: Enhance Convex user sync for GitHub and Google auth
- Updated ConvexUserSync to handle both GitHub and Google OAuth providers
- Fixed SSR/build-time issues by deferring Convex hooks to client-side only
- Split sync logic into separate client component to avoid hydration errors
- Users are now automatically persisted to Convex DB on auth
- Chat history and README generation will use Convex as the database
